Access Control and File Upload Added - Final Commit

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -51,10 +62,31 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
+        // Handle File Upload
+        if ($request->hasFile('cover_image')) {
+            // Get Filename with the extension
 
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            // Get Just Filename
+            $filname = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            
+            // Get Just Ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            
+            // Filename to store
+            $fileNameToStore = $filname.'_'.time().'.'.$extension;
+
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         // Create Post
 
@@ -62,6 +94,7 @@ class PostsController extends Controller
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -89,6 +122,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit', compact('post'));
     }
 
@@ -106,11 +145,34 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
+         // Handle File Upload
+         if ($request->hasFile('cover_image')) {
+            // Get Filename with the extension
+
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            // Get Just Filename
+            $filname = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            
+            // Get Just Ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            
+            // Filename to store
+            $fileNameToStore = $filname.'_'.time().'.'.$extension;
+
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }
+
         // Create Post
 
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -125,8 +187,18 @@ class PostsController extends Controller
     public function destroy(Request $request, $id)
     {
         $post = Post::find($id);
-        $post->delete();
 
+         // Check for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }   
+
+        $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
 
     }
